시설 건설 시간과 비용 밸런스 재조정

+++ java-common/src/main/java/org/duckdns/hjow/colonization/elements/facilities/AbstractFacility.java
@@ -541,7 +541,7 @@ public abstract class AbstractFacility implements Facility {
     }
     
     public static Integer getFacilityBuildingCycle() {
-        return new Integer(1200);
+        return new Integer(120);
     }
     
     public static Long getTechNeeded() {
