+++ java-common/src/main/java/org/duckdns/hjow/colonization/elements/facilities/Factory.java
@@ -350,7 +350,7 @@ public abstract class Factory extends AbstractFacility implements Storage {
     }
     
     public static Integer getFacilityBuildingCycle() {
-        return new Integer(1200);
+        return new Integer(120);
     }
     
     public static int getUniqueFacilityGrade() {
