+++ java-common/src/main/java/org/duckdns/hjow/colonization/elements/facilities/ResearchCenter.java
@@ -165,7 +165,7 @@ public abstract class ResearchCenter extends AbstractFacility {
     }
     
     public static Integer getFacilityBuildingCycle() {
-        return new Integer(1200);
+        return new Integer(120);
     }
     
     public static int getUniqueFacilityGrade() {
