+++ java-defaultpack/src/main/java/org/duckdns/hjow/colonization/elements/facilities/ArchitectOffice.java
@@ -120,7 +120,7 @@ public class ArchitectOffice extends AbstractFacility {
     }
     
     public static Integer getFacilityBuildingCycle() {
-        return new Integer(1200);
+        return new Integer(60);
     }
     
     public static Long getTechNeeded() {
