+++ java-defaultpack/src/main/java/org/duckdns/hjow/colonization/elements/facilities/MagneticLevitationMetroStation.java
@@ -63,7 +63,7 @@ public class MagneticLevitationMetroStation extends TransportStation {
     }
     
     public static Integer getFacilityBuildingCycle() {
-        return new Integer(1200);
+        return new Integer(300);
     }
     
     public static int getUniqueFacilityGrade() {
