+++ java-defaultpack/src/main/java/org/duckdns/hjow/colonization/elements/facilities/Restaurant.java
@@ -308,7 +308,7 @@ public class Restaurant extends AbstractFacility implements ServiceFacility, Sto
     }
     
     public static Integer getFacilityBuildingCycle() {
-        return new Integer(180);
+        return new Integer(30);
     }
     
     public static int getUniqueFacilityGrade() {
