+++ java-defaultpack/src/main/java/org/duckdns/hjow/colonization/elements/facilities/School.java
@@ -226,7 +226,7 @@ public abstract class School extends AbstractFacility {
     }
     
     public static Integer getFacilityBuildingCycle() {
-        return new Integer(180);
+        return new Integer(60);
     }
     
     public static int getUniqueFacilityGrade() {
