+++ java-defaultpack/src/main/java/org/duckdns/hjow/colonization/elements/facilities/SmallFactory.java
@@ -62,7 +62,7 @@ public class SmallFactory extends Factory {
     }
     
     public static Integer getFacilityBuildingCycle() {
-        return new Integer(1200);
+        return new Integer(120);
     }
     
     public static int getUniqueFacilityGrade() {
